SCALA 17 - Moving the ScalaCollection file to commoncollections package (#98)
