+++ core-scala-collections/src/main/com/baeldung/scala/commoncollections/ScalaCollections.scala
@@ -1,4 +1,4 @@
-package com.baeldung.scala
+package com.baeldung.scala.commoncollections
 
 import scala.collection.mutable
